Set up virtualenv and install custom SDK only if SDK appears in arvados/sdk/python -- ignore sdk/python.
This means scripts that are run straight from the arvados tree will
use whatever SDK is installed on the compute node, rather than
installing the SDK and dependencies into a virtualenv.

If you want your scripts to be in [a fork of] the arvados tree rather
than bringing in arvados as a submodule, *and* you want to pin jobs to
a specific SDK version, add a symbolic link in your source tree
(/arvados -> .).