Correct parameterized input collection reference
Attempting to follow the tutorial I was blocked because the UUID for the alternate input supplied as a command line parameter was incomplete. This can be easily overlooked because in fact the final filtered output is a zero-length file in either case: it's just that currently it is zero-length only because its input is empty!