api: Improve compute_node_domain handling.
The default value was causing test failures on my machine because the
ERB evaluated to an empty string.  And we want to avoid this causing
any DNS lookups during tests.  Now we make the default false (in line
with our other compute_node_* settings), and suggest a change in
application.yml.example to demonstrate how it should be set.

I also fixed up the example about compute_node_nameservers while I was
at it.

This closes #2351.