workbench: Add developer docs.
This adds a new README with information that's useful for people who
want to develop Workbench.  Right now it's just information about how
to run and write tests.  We can add to this as we go.

Some of this information was pulled from the installation guide.  I
also deleted Rails' stock README while I was at it.