workbench: Add basic smoke test.
This visits all the link on a user's front page and makes sure they
all return 200.

It's easy to extend this to crawl the entire site, checking for
success.  I've included the code for that.  We need to squash some
bugs and check the timing before we start doing that, though.

This overlaps api_client_authorizations_test, so I deleted that.