doc: Use $USER in tutorial JSON.
The tutorial encourages users to copy and paste command text.
However, most of the JSON needs to have a user-specific value for
"repository", which is easy to overlook.  Since we're telling users to
save JSON with @cat@, we can take advantage of shell variable
expansion to make the right thing happen automatically.

I added accompanying notes to help explain what's going on for people
who aren't copying instructions so literally.

Conflicts:
	doc/user/tutorials/tutorial-firstscript.html.textile.liquid
	doc/user/tutorials/tutorial-new-pipeline.html.textile.liquid