Accept minimum_script_version, excluded_script_versions, and find_or_create flag as parameters to create, instead of treating them as if they were attributes of the Job resource.
Accept find_or_create as a positive flag, rather than no_reuse as a
negative flag.

Behave as requested by the find_or_create flag even if
nondeterministic=true in the submitted job: i.e., a client can use the
find_or_create feature when submitting a job which itself will never
be reused. (It's up to the client to decide whether this is ever
useful.)

For now, both old and new flag names/locations are accepted and
supplied by API server and arv-run-pipeline-instance respectively.