Refactor AdminNotifier.
The old implementation was elegant, but not super useful in the
situation where you want to send different e-mails from a method.

Due to the way ActiveMailer is implemented, Rails would reuse the
Mail::Message object for sending the messages, which makes this
impossible to test.

So... I threw out the fancy solution in favor of something that also
works and can be tested fully.

I also added a number of tests.