workbench: File downloads include reader tokens.
This makes it possible to download an entire collection by pointing
`wget -r` at a collection page with a reader_tokens parameter.

wget respects robots.txt, so this required a finer-grained approach to
shooing away robots.  We use a meta tag to mark everything as NOINDEX,
and implement NOFOLLOW on everything except Collections pages.