2759: Workbench shouldn't try to make links from Link names.
Because the name attribute can include arbitrary text, it might
resemble an Arvados object UUID.  If that happens, the
_arvados_attr_value partial will try to make a link from it.  All
sorts of weirdness could happen as a result (our specific bug was a
bad URL fed to Workbench's API client), so don't even try.  Just
render the name directly as editable text.

Closes #2759.