2044: Non-admins get all users' basic info from index API.
This will support a Workbench feature to let users at the same site
share projects with each other.  Adding kind to index requests with
select parameters is necessary to help Workbench understand the
result.

There was extensive discussion on IRC about whether or not e-mail
addresses should be included in this information.  We toyed with ideas
like providing an e-mail address checksum, so that you could find
exact but not partial matches.  Ultimately I decided that none of
those measures were worth the hassle, because the domain of addresses
at any given site would be small enough that they would easily be
discoverable through brute force.