arv-run-pipeline-instance doesn't submit nil Job parameters.
This is the same principle that we already take with POST parameters,
applied to the body of the Job itself.  It immediately prevents
sending nil runtime_constraints, which the API server doesn't accept.
It should also mean that any error messages passed on from the API
server are more likely to be helpful.

No issue # - reported by Abram on IRC.