3201: Only Projects can be moved to "My Projects."
Before this commit, it was possible to move any object that goes in
folders to "My Projects."  If you did, it would be removed from any
project, and owned by the user directly.  This is non-obvious.  Data
should be removed from projects by using the trash icon on the project
page instead.

Closes #3201.