3252: Include README in Python source manifest.
The real issue with our Python package builds is that this file isn't
being included, which causes setup.py to fail when it's run from a
source tarfile.  It *is* being included on my box, but that seems to
be because of extra information from git, or lingering information in
my .egg-info directory.  This change should make sure README is
included when a clean sdist is prepared.

This reverts commit e16c1ecc8097fd586e86fc90206836000b5f5004,
since that wasn't the real bugfix.  Refs #3252.