3531: Move the tab_line_buttons sections out of the tab pane partials into the appropriate "index" and "show" pages.
Until now, we have been predicting which tab pane will be rendered
with the initial page view, and putting content_for :tab_line_buttons
in that pane. But the tab_line_buttons aren't really in that tab pane,
so they really don't deserve to be in that partial.

This worked as long as the initial tabs were being rendered in a
different way than subsequent tab loads. Now that we render all tab
pane contents using the same helper method, we have to clean this up.