3457: Revamp Python SDK and tool logging.
The primary goal of this commit is to introduce a dedicated logging
object to the arvados module, which we configure to our
specifications, and use it consistently throughout the SDK and tools.
This lets us handle our logging configuration consistently in one
place, and avoid interfering with logging in SDK users.  I made some
minor style changes to logging calls while I was at it.