2800: Remove global state from KeepClient.
This commit makes it possible to build and use a KeepClient that isn't
influenced by changes in outside state.  Changing the KeepClient based
on global state has been pushed up to the simple Keep class.

This commit changes the tests as little as possible to demonstrate
backward compatibility.  Refactoring the tests to build KeepClients
directly will come in the future.