3414: adding PullHandler and Replicator
Added PullHandler and a "replicator" package to handle "PUT /pull" requests.

PUT /pull requests are routed to PullHandler, which authenticates the
request and validates the JSON in the request body. Valid requests are
sent to the replicator.

The Keepstore replicator runs a goroutine which repeatedly listens on
its input channel for a new pull list.

TestPullHandler tests each combination of: request from superuser;
request from ordinary user; properly formatted pull request; improperly
formatted pull request.  It checks the state of the replicator when done
to make sure that it has the expected number of pull requests.

3414: add replicator.