3720: Refactor manifest parsing from API server to Ruby SDK.
To date, the API server has been parsing manifests and returning
parsed information in the files attribute.  That's convenient, but
causes performance problems in large Collections.  Profiling indicates
that most of the API server's response time is spent in the
as_api_response method; rendering the same information twice really
hurts.

This commits removes the API server's need to always parse manifests,
and moves the parsing code to the Ruby SDK.  Both the API server and
Workbench use this Gem to parse manifests on an as-needed basis.
This, combined with improvements to the new Keep::Manifest class, lays
the groundwork for future performance gains.