3704: Treat project filters consistently in Workbench chooser.
Before this, a chooser that was loaded on a specific project would
return empty results if the user selected a different project.  This
is because it would search for items with two different owner_uuids,
both the project selected at load time, and the project the user
selected later.  This patch fixes that by separating the project
filter from other filters, and preseeding it in the same place where
it's updated by the project selection AJAX.  Closes #3704, #3778.