Install & test Ruby SDK, then use that version for other tests.
This commit gives our Ruby SDK the same treatment as other parts of
our build process: we build and install it early, then test it.  We
use that built version to run other tests, to make sure all the
components are synced up throughout the run.

This required turning off `bundle install --deployment`, because
Bundler will not use locally installed Gems in that mode.  This does
make our build process a little less like what we use in production,
but the benefits of consistent testing and tightening the build
loop (you can update a Gem and dependent Gemfiles in one push) seem
worth that cost.

Refs #3720.