3894: gem version numbers reflect latest commit in the subdirectory where the gem content comes from.
Even if there are more recent commits elsewhere in the source tree,
those changes won't be reflected in the package so there's no need to
increment the package version.

In addition to avoiding the proliferation of packages with no changes
other than the version number, this makes it possible to change an
SDK, predict the gem version that the CI server will build, refer to
that version in a Gemfile elsewhere in the source tree, and push both
changes at once -- without time travel.