4189: Workbench lets users with write permission rename objects in projects.
The key part of this commit are the changes to ApplicationHelper.
There's also some refactoring to more consistently answer questions
about whether or not an attribute is editable.

There are also some tests in here that I wrote while I was tracking
down the problem.  They seem like a useful addition.