4015: simplify integration tests
Responding to code review in #4015-8:

We can (and should) test that the collection portable_data_hash is
recorded in individual tests without adding explicit new tests or
fixtures.

Updated test 'Create and run a pipeline' and the helper
create_and_run_pipeline_in_aproject to check that the collection
portable_data_hash, uuid and name have been recorded correctly in the
Advanced tab.