Fix race condition in Node Manager shutdown tests.
ComputeNodeMonitorActor considers a shutdown when it's initialized.
Blocking on the subscription we send ensures that message is fully
processed before the tests continue.  This is important because
several of them change the shutdown state immediately after, and that
might be seen by the internal shutdown consideration without this.

No issue #, but see
<https://ci.curoverse.com/job/arvados-api-server/1027/console>.