4591: Websockets server fetches fewer logs at a time.
Most of the out of memory errors we're seeing happen in the PostgreSQL
driver, which runs out of space to store results.  Because Log records
are relatively large (holding two other records as JSON text),
fetching fewer in a batch should noticeably improve memory use.  I
don't expect this to end the crashing, though—it seems like the
Websockets server grows large for a variety of reasons.  Hopefully
this change will help make some of the others clearer.