4836: Fix infinite scroll loading in background tabs.
If an infinite scroller is in the first tab of a show page, but the
user is going to a different tab, we'll queue up the first event
to load data for the container, but when it fires the container won't
be visible so it will decline to load anything.  Then you can only get
data to load if you resize the window.

Checking for container visibility was done in
ffa64c065ac8b28849395cf3e89fe93814252473, when maybe_load_more_content
would run on multiple containers.  That was obsoleted in
4ad1947da490a998d68380b1520c4f95b1784676, when the function was
attached as an event handler to each container individually.  Checking
for visibility is obsolete and can safely be removed.