4967: crunch-job unmounts old Keep mounts more robustly.
The previous version looked for mounts under $JOB_WORK before
unmounting anything.  We long ago started mounting under $TASK_WORK,
so unmounting hasn't been working for a while.  This version unmounts
every FUSE mount with "keep" in the path, so it will be much more
flexible.

I also changed the cleanup to use bash -e, so the logs are more likely
to indicate if there was trouble during the cleanup step.