5037: Fix race condition in PySDK Websockets test.
This might just be bad luck, but for some reason when testing Tom's
changes I ran into a situation where this test failed because it
started processing the Websocket event about the new human before it
assigned self.h, so the assertion failed.  Fix the race condition by
asserting against a thread-aware Queue.