5175: Do not add +sha1 tag to Python packages ever.
PEP 440 is clear that the local version tag is for distributors who
are patching an upstream release.  That's not us; we are the upstream
release.  "Local version identifiers SHOULD NOT be used for upstream
projects," so just don't do it.

This also lets us sidestep the issue of trying to parse sys.argv
behind setuptools' back.

Developers who want to keep their development builds distinct should
tag packages themselves with -b .devN.