5313: Node Manager has cloud-specific logic to get node FQDNs.
On AWS, we put compute nodes' FQDN in the name field.  On GCE, we
can't do that: it can't contain dots.  Add a node_fqdn classmethod to
cloud drivers to get a node's FQDN from the right place, and use that
method when deciding whether or not to sync a node.