4253: Users can manage their own repositories.
This commit allows users to create their own repositories, as long as
the repository name starts with their own username.

To support this change, we've modified our Gitolite setup to store
repositories primarily by UUID, with a name alias for easier
checkout.  fetch_url and push_url become generated attributes
accordingly.  This makes it easier to rename the repository later and
allow checkouts to continue to work.