5352: crunch-dispatch treats node allocation failure as temporary.
Imagine a scenario where multiple crunch-dispatch processes are
sitting idle, then suddenly a new job appears in the queue.  They will
all race to dispatch the job.  When this happens, we frequently see
that salloc fails for most of them, because they all requested the
same node(s) and only the winner will get them.  crunch-dispatch has
no way to know the exit code "came from" salloc and not crunch-job,
and so marks the job failed.

This patch sets the SLURM_EXIT_IMMEDIATE environment variable to make
salloc use exit code 75 when the allocation fails.  crunch-dispatch
already recognizes this exit code as a temporary failure, and will
leave the Arvados job record unchanged.  Refer to salloc(1) and the
long comment in Dispatch#reap_children.