Tighten up DNS check in arvdock.
I got arvdock stuck in an infinite loop because I had the right
nameserver line in /etc/resolv.conf, except it was commented out, so
arvdock kept checking for a name that it would never find.  Use awk to
find and check the first functional nameserver line instead.

No issue #.