5490: Test API server handling of git hash-like branch names.
If you have a branch that looks like a commit hash, different git subcommands do different things.  After extended discussion on IRC, we've decided we'd like the rules to be:

* If a 40-character hex string references a commit, resolve it to that
  commit.
* In other cases of ambiguity, prefer branch names over short commit
  hashes.

Fortunately, this corresponds to the behavior of `git rev-list` and
our existing implementation.  This commit codifies our desired
behavior with tests.