5554: Workbench links finished job logs from pipeline instance log tab.
This is an affordance to help users find logs after they're removed
from the API server's logs table.  The display is not very great.
We're punting improvements until Workbench generally has better
capabilities to display large logs.

The layout change necessitated updates to the anonymous access tests.
While I was in there, I strengthened the assertions by putting
assertions that check a particular element has loaded before asserting
the nonexistence of other elements.  This revealed the JavaScript
issue mentioned in the new comment: the old version of the test was
doing a single `assert_no_text` after clicking a job's log tab, which
only confirmed that we navigated away.  We already test that the log
tab exists and is not disabled in this case, so I consider that
sufficient for now.